+++ src/SMAPI/IModInfo.cs
@@ -1,0 +1,9 @@
+namespace StardewModdingAPI
+{
+    /// <summary>Metadata for a loaded mod.</summary>
+    public interface IModInfo
+    {
+        /// <summary>The mod manifest.</summary>
+        IManifest Manifest { get; }
+    }
+}